+++ host/CMakeLists.txt
@@ -3,8 +3,8 @@ target_include_directories(host INTERFACE ${CMAKE_CURRENT_LIST_DIR}/include)
 target_link_libraries(
     host
     INTERFACE
-    ipc_data_reader
-    ipc_data_writer
+    data_reader
+    data_writer
     service
 )
