+++ ipc/CMakeLists.txt
@@ -1,5 +1,5 @@
-add_library(ipc_data_reader INTERFACE)
-target_include_directories(ipc_data_reader INTERFACE ${CMAKE_CURRENT_LIST_DIR}/include)
+add_library(data_reader INTERFACE)
+target_include_directories(data_reader INTERFACE ${CMAKE_CURRENT_LIST_DIR}/include)
 
-add_library(ipc_data_writer INTERFACE)
-target_include_directories(ipc_data_writer INTERFACE ${CMAKE_CURRENT_LIST_DIR}/include)
+add_library(data_writer INTERFACE)
+target_include_directories(data_writer INTERFACE ${CMAKE_CURRENT_LIST_DIR}/include)
